backend/requirements.txt
fastapi
uvicorn
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
pydantic
openai
python-dotenv
